Add files via upload
Manual for running the code
The code uses an open cv, a computer vision, and NumPy, an array manipulation library for
this task. To run the code, at first, the users are required to open the code folder in any IDE of
their preference. After that, it is necessary to install all the dependencies. Below is the list
with the command to install these dependencies:
Py -m pip install pylsd
Py -m pip install python-opencv
Py -m pip install numpy
These are three basic dependencies that must be installed to run the code. After installing
these dependencies, the user should type py -m python run.py, which will run the program.
Furthermore, it is to be noted that the input image must be saved as image.jpg in the image
folder under the code folder.